The reducer has been split into several different
