+++ src/index.js
@@ -4,7 +4,7 @@ import {Provider} from 'react-redux';
 
 import '../src/scss/index.scss';
 import App from './App';
-import store from './store/store';
+import store from './redux/store';
 
 ReactDOM.render(
   <React.StrictMode>
