+++ src/redux/store.js
@@ -1,23 +1,7 @@
 import { applyMiddleware, compose, createStore } from 'redux';
 import thunk from 'redux-thunk';
 
-import shopReducer from './reducers';
-
-const user = localStorage.getItem('user') ? JSON.parse(localStorage.getItem('user')) : {};
-
-const initialState = {
-  user,
-  books: {
-    data: [],
-    isLoading: false,
-    error: null,
-  },
-  cart: {
-    books: [],
-    totalCount: 0,
-    totalPrice: 0,
-  },
-};
+import rootReducer from './reducers';
 
 const composeEnhancers = typeof window === 'object'
   && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
@@ -28,8 +12,7 @@ const enhancer = composeEnhancers(
 );
 
 const store = createStore(
-  shopReducer,
-  initialState,
+  rootReducer,
   enhancer,
 );
